Fix conversions

+++ src/libs/lang/include/lang/mem.h
@@ -1,6 +1,7 @@
 #if !defined(LANG_MEM_H)
 #define LANG_MEM_H
 
+#include <stddef.h>
 #include "except.h"
 
 /**
@@ -29,7 +30,7 @@ extern const Except_T Mem_Failed;
  *
  * @throw `Mem_Failed` if NDEBUG is not defined
  */
-extern void* Mem_alloc(long nbytes, const char* file, int line);
+extern void* Mem_alloc(size_t nbytes, const char* file, int line);
 /**
  * Allocates space for an array of `count` elements, each occupying `nbytes`
  * bytes, and returns a pointer to the first element. It is a checked run-time
@@ -40,7 +41,7 @@ extern void* Mem_alloc(long nbytes, const char* file, int line);
  *
  * @throw `Mem_Failed` if NDEBUG is not defined
  */
-extern void* Mem_calloc(long count, long nbytes, const char* file, int line);
+extern void* Mem_calloc(size_t count, size_t nbytes, const char* file, int line);
 
 /**
  * De-allocates `ptr`, if `ptr` is non NULL. It is a unchecked run-time error
@@ -64,7 +65,7 @@ extern void  Mem_free(void* ptr, const char* file, int line);
  *
  * @throw `Mem_Failed` if NDEBUG is not defined
  */
-extern void* Mem_resize(void* ptr, long nbytes, const char* file, int line);
+extern void* Mem_resize(void* ptr, size_t nbytes, const char* file, int line);
 
 #define ALLOC(nbytes)          Mem_alloc((nbytes), __FILE__, __LINE__)
 #define CALLOC(count, nbytes)  Mem_calloc((count), (nbytes), __FILE__, __LINE__)
